src/api/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
cattrs==24.1.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.1
Deprecated==1.2.14
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.2
flatbuffers==24.3.25
h11==0.14.0
idna==3.10
importlib_resources==6.4.5
limits==3.13.0
motor==3.6.0
numpy==2.1.2
openmeteo_requests==1.3.0
openmeteo_sdk==1.17.0
packaging==24.1
pandas==2.2.3
passlib==1.7.4
platformdirs==4.3.6
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pymongo==4.9.2
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytz==2024.2
redis==5.1.1
requests==2.32.3
requests-cache==1.2.1
retry-requests==2.0.0
shapely==2.0.6
six==1.16.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.40.0
typing_extensions==4.12.2
tzdata==2024.2
url-normalize==1.4.3
urllib3==2.2.3
uvicorn==0.32.0
wrapt==1.16.0

httpx==0.27.0
prometheus-client==0.20.0
opentelemetry-distro==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-logging==0.45b0
opentelemetry-exporter-otlp==1.24.0